Move theme picker from debug window to settings
